requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.24
contextlib2==0.5.4
coreapi==2.1.1
decorator==4.0.10
Django==1.10.4
django-celery==3.1.17
django-environ==0.4.1
django-extensions==1.7.5
django-filter==0.15.3
django-mailgun==0.9.1
django-rest-swagger==2.1.0
djangorestframework==3.5.3
djangorestframework-bulk==0.2.1
djangorestframework-filters==0.9.1
djangorestframework-jwt==1.9.0
flake8==3.2.1
ipdb==0.10.1
ipython==5.1.0
ipython-genutils==0.1.0
itypes==1.1.0
kombu==3.0.37
mccabe==0.5.3
openapi-codec==1.2.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.9
psycopg2==2.6.2
ptyprocess==0.5.1
pycodestyle==2.2.0
pyflakes==1.3.0
Pygments==2.1.3
PyJWT==1.4.2
python-http-client==2.2.1
pytz==2016.10
raven==5.32.0
redis==2.10.5
requests==2.12.4
sendgrid==3.6.3
sendgrid-django==4.0.4
simplegeneric==0.8.1
simplejson==3.10.0
six==1.10.0
traitlets==4.3.1
uritemplate==3.0.0
vine==1.1.3
wcwidth==0.1.7
Werkzeug==0.11.11
